Refatoração da camada de services Separação da camada de services em diversos serviços para organização de código
